DataBase: Tables can have data in multiple chunks
The row data can be split between multiple chunks in any order,
the 'index' value of the chunk being used to sort them.

When appending rows, as we can only append the last (active) one,
if our row data is not active. Then we can just end the last one
and create a new row data (RD) chunk attached to that table.